% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitCCBPs.R
\docType{data}
\name{CCBPs}
\alias{CCBPs}
\title{complete cases (main part of the data) for the jasa paper}
\format{
a data frame: V1=ID, V2=SBP, V3=DBP, V4-V64: Pollutants(PCBs,POPs, and Metals, total=62)..
}
\source{
NHANES plus imputation
}
\usage{
CCBPs
}
\description{
NHANES Data with 64 variables and 833 observations
}
\keyword{datasets}
