% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIVE.R
\name{CHIVE}
\alias{CHIVE}
\title{Estimate the explained variation in high-dimensional linear model using the method CHIVE proposed
by Cai and Guo (2020)}
\usage{
CHIVE(y, x, xext = NULL, alpha = 0.05)
}
\arguments{
\item{y}{outcome: a vector of length n.}

\item{x}{covariates: a matrix of nxp dimension.}

\item{xext}{supplementary covariates of Nxp dimension.}

\item{alpha}{confidence level is 100*(1-alpha)\%}
}
\value{
Estimate of the proportion of explained variation, variance estimates,
and confidence intervals.
}
\description{
This function estimates:
(1). the proportion of the explained variation
(2). the explained variation
by covariates in a linear model assuming the covariates have sparse effects.
}
\details{
Both point estimate and confidence intervals are computed.
}
\examples{
\dontrun{CHIVE(y,x,lam=0.05)}

}
\references{
Cai, T. T. Guo, Z. (2020).  Semisupervised inference for explained
variance in high dimensional linear regression and its applications,
Journal of Royal Statistical Society, Ser. B., 82, 391-419.
}
