% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decor.R
\name{decor}
\alias{decor}
\title{Decorrelate the data matrix to be used in the heritability estimation}
\usage{
decor(dat, decorate = 0.5, inter = 0)
}
\arguments{
\item{dat}{is a nxp data matrix}

\item{decorate}{the block size in term of a fraction of sample size n.}

\item{inter}{0 means do block wise once, inter=1 means do another interlanced decorrelation}
}
\value{
precision matrix estimate
}
\description{
This function decorrelate the data matrix to make it approximately uncorrelated.
}
\details{
It does not require to know the correlation structure and can handle $n<=p$.
The idea is to estimate the inverse of the blockwised covariance matrix rather than
the inverse of the covariance matrix which may not be feasible when n<=p.
}
\references{
To be added.
}
