% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenPrism.R
\name{EigenPrism}
\alias{EigenPrism}
\title{EigenPrism procedure for estimating and generating confidence intervals}
\usage{
EigenPrism(
  y,
  X,
  invsqrtSig = NULL,
  alpha = c(0.01, 0.05, 0.1),
  target = "beta2",
  zero.ind = c(),
  diagnostics = T
)
}
\arguments{
\item{y}{response vector of length n}

\item{X}{n by p design matrix. Columns are automatically centered and scaled to variance 1, and they cannot include
one for the intercept terms.}

\item{invsqrtSig}{if columns of X are not independent, p by p positive definite matrix which is the square-root
of the inverse of Sig, where Sig is the \emph{correlation} matrix of the X. Default is identity}

\item{alpha}{significance level for confidence interval. Default is 0.05}

\item{target}{target of estimation/inference: options include "beta2", "sigma2", or "heritability"}

\item{zero.ind}{vector of which indices of the weight vector w to constrain to zero. Default is none}

\item{diagnostics}{boolean variable indicating whether to generate the diagnostic plots for V_i,
lambda_i, and w_i. Default is \code{TRUE}.}
}
\value{
estimate: unbiased estimate of the target (for heritability, only approximately unbiased);
CI: 100*(1-alpha)\% confidence interval for target
}
\description{
This function implements the EigenPrism procedure for estimating and generating
confidence intervals for variance components in high-dimensional linear model.
}
\details{
Details of this function, some explanations of the input arguments...
}
\examples{
\dontrun{EigenPrism(y, x)}

}
\references{
Janson, L., Barber, R. F., Candes, E. (2017). EigenPrism: inference for high-dimensional signal-to-noise ratios.
\emph{Journal of Royal Statistical Society, Ser. B.}, \strong{79}, 1037-1065.

Lucas Janson. \url{http://lucasjanson.fas.harvard.edu/code/EigenPrism.R}.
}
