% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenPrismFull.R
\name{EigenPrismFull}
\alias{EigenPrismFull}
\title{EigenPrism procedure integrating the low and high dimensional cases}
\usage{
EigenPrismFull(y, x, alpha = c(0.01, 0.05, 0.1))
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{sigma}{significance level}
}
\value{
Estimate and CI
}
\description{
One paragraph describing this function...
}
\details{
Details of this function...
}
\examples{
\dontrun{...}

}
\references{
reference 1 here...
}
