% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{example}
\alias{example}
\title{Example data for method illustration}
\format{
a data frame with 3262 records and 107 variables
}
\source{
NHANES plus imputation
}
\usage{
example
}
\description{
NHANES Data with 7 outcomes, 12 possible confounders and 88 exposures
There are 3262 records.
}
\keyword{datasets}
