% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCTAREML.R
\name{GCTAREML}
\alias{GCTAREML}
\title{This function implement the maximum likelihood approach with variance estimated
by inverse of information matrix assuming sigma^2_Y=1.}
\usage{
GCTAREML(y, x, alpha = c(0.05), lam = 1, niter = 100, eps = 1e-06)
}
\arguments{
\item{y}{outcome: a vector of length n.}

\item{x}{covariates: a matrix of nxp dimension.}

\item{alpha}{a vector of type I errors used to generate (1-alpha)confidence intervals.}

\item{lam}{initial value}

\item{niter}{the number of iterations for finding the signal noise ratio}

\item{eps}{the convergence criterion for the iteration}
}
\value{
Estimate of proportion of the explained variation, variance estimates,
and the corresponding confidence intervals.
}
\description{
This is the reml for the high dimensional linear model assuming known sigma_Y^2=1
}
\details{
This method assume the independent covariates with fixed effects but can be
equivalently treated as random effects
}
\examples{
\dontrun{GCTAREML(y,x)}

}
\references{
Restricted maximum likelihood estimator of the random effects model
}
