% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRE.R
\name{GRE}
\alias{GRE}
\title{This function implement the generalized random effects model approach.}
\usage{
GRE(y, x, alpha = c(0.05))
}
\arguments{
\item{y}{outcome: a vector of length n.}

\item{x}{covariates: a matrix of nxp dimension.}

\item{alpha}{a vector of type I errors used to generate (1-alpha)confidence intervals.}
}
\value{
Estimate of proportion of the explained variation, variance estimates,
and the corresponding confidence intervals.
}
\description{
This function uses least-square estimates in computing the proportion of the explained variation.
It is simply just the least-square approach to estimate heritability.The general GRE assumes
covariates are blockwise independent. This function can be applied to one block. Variance estimate
is good for normal outcome only.
}
\details{
This method works only for the case n>p. It uses the least-square approach
for the estimation. Covariates are allowed to be correlated.
}
\examples{
\dontrun{GRE(y,x)}

}
\references{
Hou et al (2019).  Accurate estimation of SNP-heritability from
biobank-scale data irrespective of genetic architecture. NAture GeNeticS, 51, 1244–1251.
}
