% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitIMPallcov.R
\docType{data}
\name{IMPallcov}
\alias{IMPallcov}
\title{six copies of imputed data for the jasa paper}
\format{
a data frame: V1=ID, V2=imputation index, V3-9: confounders
V10-V71: Pollutants(PCBs,POPs, and Metals, total=62).
}
\source{
NHANES plus imputation
}
\usage{
IMPallcov
}
\description{
NHANES Data with 71 variables and 28986 observations
}
\keyword{datasets}
