% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jasaSimulation.R
\name{jasaSimulation}
\alias{jasaSimulation}
\title{Simulation studies for JASA submission}
\usage{
jasaSimulation(
  cindep = "T",
  sparse = "F",
  n = 400,
  N = 200,
  p = 200,
  p1 = 100,
  effectconst = 1,
  climeconst = 1,
  powx = 1,
  powy = 1,
  xsig = 1,
  errsig = 1,
  method = "our",
  rho = 0.9,
  nrep = 1000,
  rept = 1000
)
}
\arguments{
\item{cindep}{determine if covariates are independent: cindep="T" if independent,
cindep="F" if dependent.}

\item{sparse}{determine whether effects are sparse: spare="T" if effects are sparse,
sparse="F" if effects are dense}

\item{n}{sample size of main data, e.g., n=400.}

\item{N}{sample size for supplementary data, e.g., N=200, it is set to N=p in the simulation.}

\item{p}{dimension of the covariates, for example, p=200,}

\item{p1}{subdimension of covariates having nonzero effects p1<=p, e.g. p1=100 when p=200.
in the simulation of sparse effects, p1=4.}

\item{effectconst}{the constant for determining regression coefficients. it needs to be adjusted to achieve R2 wanted.}

\item{climeconst}{the const used in adjusting selection in CLIME constaints in R-package flare.}

\item{powx}{powertansformation of normal covariates, powx=1 means normal, powx=2 means chi-square}

\item{powy}{powertansformation of normal random error, powx=1 means normal, powx=2 means chi-square}

\item{xsig}{standard deviation of covariates}

\item{errsig}{standard deviation of random error.}

\item{method}{method for dependence generate: method="our" means our approach, otherwise Cai and Guo's approach.}

\item{rho}{correlation coefficient in the AR(1) covariate dependence model. This has effects only when method is not "our".}

\item{nrep}{number of replicates in simulation sample for variance estimation nrep=1000 by default.}

\item{rept}{number of replicates in simulation rept=1000 by default.}
}
\value{
Estimate of the proportion of the explained variation and confidence intervals for the proportion.
}
\description{
Simulation studies for JASA submission
}
\examples{
\dontrun{jasaSIMULATION(cindep="T",sparse="F", p=200, p1=100, n=400,N=200,
                                  powx=1,powy=1,xsig=1,errsig=1,rho=0.9,rept=100,nrep=1000)}

}
