% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lezhong.R
\name{lezhong}
\alias{lezhong}
\title{Estimate the precision matrix when the network structure is known}
\usage{
lezhong(dat, network)
}
\arguments{
\item{dat}{is a nxp data matrix}

\item{network}{is the fixed sparse struct of the precision matrix given}
}
\value{
precision matrix estimate
}
\description{
This function find the precision matrix estimate by the method of Le and Zhong (2022)
}
\details{
This approach works for sparse precision matrix
}
\references{
Le, T. M. and Zhong, P. S. (2022) High-dimensional precision matrix estimation with a known graphical structure. Stat, 11, e424.
}
