% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{makecorb}
\alias{makecorb}
\title{generate a square root of a correlation matrix by specifying correlation matrix}
\usage{
makecorb(rho, p, fix = TRUE)
}
\arguments{
\item{rho}{parameter controlling the correlation}

\item{p}{dimensional of the matrix}

\item{fix}{the correlation is a constant if fix=TRUE, Otherwise, it is rho^|i-j|}
}
\value{
corr correlation matrix
}
\description{
This function generate data follows linear model based on parameter input.
}
