% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{makedat}
\alias{makedat}
\title{Generate data for simulation studies.}
\usage{
makedat(n, p, beta, xsig, errsig, powx, powy, sqrtsig)
}
\arguments{
\item{n}{sample size}

\item{p}{number of covariates}

\item{beta}{regression coefficients}

\item{xsig}{standard deviation of x variables}

\item{errsig}{residual standard deviation}

\item{powx}{power for x variable transformation}

\item{powy}{power of y variable transformation}

\item{sqrtsig}{square root decomposition of correlation matrix}
}
\value{
x covariates.
y outcome variable
}
\description{
This function generate data follows linear model based on parameter input.
}
