% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{makedata}
\alias{makedata}
\title{Generate data following high-dimensional linear model.}
\usage{
makedata(n, p, beta, xsig, errsig, powx, powy, k, a = rep(0.5, k))
}
\arguments{
\item{n}{sample size}

\item{p}{number of covariates}

\item{beta}{regression coefficients}

\item{xsig}{standard deviation of x variables}

\item{errsig}{residual standard deviation}

\item{powx}{power for x variable transformation}

\item{powy}{power of y variable transformation}

\item{k}{the length of dependence in the auto-regression}

\item{a}{autoregressive coefficients of length k}
}
\value{
x covariates.
y outcome variable
}
\description{
This function generates data following a linear model where covariates follow
an auto regressive model.
}
