% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodewisereg.R
\name{nodewisereg}
\alias{nodewisereg}
\title{Approximate MLE with given precision matrix structure}
\usage{
nodewisereg(dat, network)
}
\arguments{
\item{dat}{is a nxp data matrix}

\item{network}{is the fixed sparse struct of the precision matrix given}
}
\value{
precision matrix estimate
}
\description{
This function find the MLE of the precision matrix approximately by individual regression
models when the sparse structure of the precision matrix is given.
}
\details{
This approach works for very sparse precision matrix
}
\references{
This program is nodewise regression for precision matrix.
}
