% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proj2}
\alias{proj2}
\title{Compute projection matrices}
\usage{
proj2(x, X)
}
\arguments{
\item{x}{input matrix of nxp dimension.}

\item{X}{input matrix of Nxp dimension.}
}
\value{
projection matrix of nxn and projection matrix of NxN based on (x, X).
}
\description{
Compute projection matrices
}
