% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitDataAnalysis.R
\name{projed}
\alias{projed}
\title{project data matrix x to the orthogonal complement of data matrix z.}
\usage{
projed(x, z)
}
\arguments{
\item{x}{matrix (nxp) to be projected.}

\item{z}{matrix (nxq) the projection is based on.}
}
\value{
the projected x.
}
\description{
This function uses linear projector (I-z(z^tz)^(-1)z^t) to project.
}
\details{
compute (I-z(z^tz)^(-1)z^t)x
}
\examples{
\dontrun{projed(x,z)}

}
