% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2ee.R
\name{R2ee}
\alias{R2ee}
\title{Estimating equation approach to the proportion of the explained variation}
\usage{
R2ee(y, x, lam = 1, niter = 1)
}
\arguments{
\item{y}{outcome: a vector of length n.}

\item{x}{covariates: a matrix of nxp dimension.}

\item{lam}{parameter for altering the weighting matrix.}

\item{niter}{number of iterations for updating the lam parameter.}
}
\value{
Estimate of the proportion of explained variation, variance estimates under normality
and non-normality assumptions, and confidence intervals under normality and non-normality assumptions.
}
\description{
This approach estimates the proportion of the explained variation in a linear model
assuming the covariates are independent.
}
\details{
Both point estimate and confidence intervals
are computed. Two set of confidence intervals under normal or non-normal error are computed.
}
\examples{
\dontrun{R2ee(y,x,lam=1.0,niter=1)}

}
\references{
Chen, H.Y. (2022). Statistical inference on explained variation in high-dimensional
linear model with dense effects. arXiv:2201.08723.

Chen, H. Y., Li, H., Argos, M., Persky, V. W., and Turyk, M. (2022). Statistical Methods
for Assessing Explained Variation of a Health Outcome by Mixture of Exposures. International Journal
of Environmental Research and Public Health.
}
