% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eeadj.R
\name{R2eeadj}
\alias{R2eeadj}
\title{Confidence intervals for the two-group estimating equation approach}
\usage{
R2eeadj(y, xa, xb, niter = 1)
}
\arguments{
\item{y}{outcome}

\item{xa}{covariates to be adjusted}

\item{xb}{covariate effects to be computed}

\item{niter}{number of iterations for updating lambda. Default is 1}
}
\value{
Output of this function...
}
\description{
One paragraph describing this function...
}
\details{
Details of this function...
}
\examples{
\dontrun{R2eeadj(y,xa,xb,niter=1)}


}
\references{
Reference 1...

Reference 2...
}
