% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eeadjPMT.R
\name{R2eeadjPMT}
\alias{R2eeadjPMT}
\title{Permutation test for the two-group estimating equation approach}
\usage{
R2eeadjPMT(y, xa, xb, niter = 5, npm = 1000, proja)
}
\arguments{
\item{y}{outcome}

\item{xa}{covariates to be adjusted}

\item{xb}{covariate effects to be computed}

\item{niter}{number of iterations for updating lambda. Default is 5}

\item{npm}{permutation sample size. Default is 1000}

\item{proja}{matrix projecting to orthogonal complement of xa}
}
\value{
Output of this function...
}
\description{
One paragraph describing this function...
}
\details{
Details of this function...
}
\examples{
\dontrun{R2eeadjPMT(y,xa,xb,niter=5,npm=1000,proja)}

}
\references{
Reference 1 here...

Reference 2 here...
}
