% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eels.R
\name{R2eels}
\alias{R2eels}
\title{Least-square approach to proportion of the explained variation}
\usage{
R2eels(y, x)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}
}
\value{
Output includes estimate of proportion of the explained variation, variance estimates
under normality and non-normality error, and the corresponding confidence intervals.
}
\description{
This function uses least-square estimates in computing the proportion of the explained variation.
}
\details{
This functions uses the least square approach and works only for n > p. But covariates are
not required to be independent.
}
\examples{
\dontrun{R2eels(y, x)}

}
\references{
Chen, H.Y. (2022). Statistical inference on explained variation in high-dimensional
linear model with dense effects. arXiv:2201.08723
}
