% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eelsPMTca.R
\name{R2eelsPMTca}
\alias{R2eelsPMTca}
\title{Conditional permutation test for the least squares approach}
\usage{
R2eelsPMTca(y, x, pa, npm = 1000)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{pa}{dimension of covariates to be adjusted}

\item{npm}{permutation sample size. Default is 1000}
}
\value{
Output of this function...
}
\description{
One paragraph describing this function...
}
\details{
Details of this function...
}
\examples{
\dontrun{R2eelsPMTca(y,x,pa,npm=1000)}

}
\references{
Reference 1 here...

Reference 2 here...
}
