% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eelsPMTcaa.R
\name{R2eelsPMTcaa}
\alias{R2eelsPMTcaa}
\title{Another conditional permutation test...}
\usage{
R2eelsPMTcaa(y, x, pa, npm = 1000)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{pa}{dimension of covariates to be adjusted}

\item{npm}{permutation sample size. Default is 1000}
}
\value{
Output of this function...
}
\description{
One paragraph descirbing this function...
}
\details{
Details of this function...
}
\examples{
\dontrun{R2eelsPMTcaa(y,x,pa,npm=1000)}

}
\references{
Reference 1 here...

Reference 2 here...
}
