% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eePMT.R
\name{R2eePMT}
\alias{R2eePMT}
\title{Permutation test for the estimating equation approach}
\usage{
R2eePMT(y, x, lam = 0.12, niter = 3, npm = 1000)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{lam}{parameter adjusting the formation of the weighting matrix. Default is 0.12}

\item{niter}{number of iterations for updating lambda. Default is 3}

\item{npm}{permutation sample size. Default is 1000}
}
\value{
The output of running this function...
}
\description{
One paragraph describing this function...
}
\details{
Details of the function...
}
\examples{
\dontrun{R2eePMT(y,x,lam=0.12,niter=3,npm=1000)}

}
\references{
reference 1 here...

reference 2 here...
}
