% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eePMTca.R
\name{R2eePMTca}
\alias{R2eePMTca}
\title{Conditional permutation test adjusting for covariates for estimating equation approach}
\usage{
R2eePMTca(y, x, pa, lam = 0.2, niter = 3, npm = 1000)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{pa}{dimension of covariates to be adjusted}

\item{lam}{paramter adjusting the format of the weighting matrix. Default is 0.2}

\item{niter}{number of iteration for updating lambda. Default is 3}

\item{npm}{Monte Carlo sample size for permutation. Default is 1000}
}
\value{
Output of this function...
}
\description{
One paragraph describing this function...
}
\details{
Details of this function...
}
\examples{
\dontrun{R2eePMTca(y,x,pa,lam=0.2,niter=3,npm=1000)}

}
\references{
Reference 1 here...

Reference 2 here...
}
