% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eePMTcab.R
\name{R2eePMTcab}
\alias{R2eePMTcab}
\title{Another conditional permutation test...}
\usage{
R2eePMTcab(y, x, pa, lam = 0.1, niter = 3, npm = 1000)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{pa}{dimension of covariates to be adjusted}

\item{lam}{parameter adjusting the format of the weighting matrix. Default is 0.1}

\item{niter}{number of iterations for updating lambda. Default is 3}

\item{npm}{Monte Carlo sample size for permutation. Default is 1000}
}
\value{
The output of this function...
}
\description{
One paragraph describing this function...
}
\details{
Details of this function...
}
\examples{
\dontrun{R2eePMTcab(y,x,pa,lam=0.1,niter=3,npm=1000)}

}
\references{
reference 1 here...

reference 2 here...
}
