% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eesd.R
\name{R2eesd}
\alias{R2eesd}
\title{confidence interval for estimating equation approach with supplementary covariate data}
\usage{
R2eesd(
  y,
  x,
  X,
  lam = 0.2,
  niter = 3,
  V = rep(0, 3),
  E = rep(0, 3),
  know = "yes",
  nrep = 1000
)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{X}{supplementary covariate data}

\item{lam}{parameter adjusting the format of the weighting matrix. Default is 0.2}

\item{niter}{number of iterations for updating lambda. Default is 3}

\item{V}{V[1:3] == V[1]=VA, V[2]=VAB, V[3]=VB}

\item{E}{E[1:3] == E[1]=EC, E[2]=ED, E[3]=EF}

\item{know}{if VA and EB are known, options include "yes" and "no". Default is "yes"}

\item{nrep}{Monte Carlo sample size for updating VA and EB}
}
\value{
Output of this function...
}
\description{
One paragraph describing this function...
}
\details{
Details of this function...
}
\examples{
\dontrun{R2eesd(y,x,X,lam=0.2,niter=3,V=rep(0,3),E=rep(0,3),know="yes",nrep=1000)}


}
\references{
reference 1 here...

reference 2 here...
}
