% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eesdPMT.R
\name{R2eesdPMT}
\alias{R2eesdPMT}
\title{Permutation test for no explained variation using the estimating equation approach with supplementary covariate data}
\usage{
R2eesdPMT(y, x, X, lam = 0.1, niter = 1, npm = 1000)
}
\arguments{
\item{y}{outcome: a vector of length n.}

\item{x}{covariates: a matrix of nxp dimension.}

\item{X}{supplementary covariates: a matrix of Nxp dimension.}

\item{lam}{parameter adjusting the formation of the weighting matrix.}

\item{niter}{number of iterations for updating lambda.}

\item{npm}{permutation sample size for simulation computation of p-value.}
}
\value{
The p-values (normal approximation, empirical estimate, and bound) of the test, estimate of proportion of
the explained variation, and simulation results.
}
\description{
This method performs permutation test by permuting the outcome values.
}
\details{
This method tests no explained variation by permuting the outcome and estimating
using the estimating equation approach with supplementary data. P-value is computed using simulation approach.
}
\examples{
\dontrun{R2eesdPMT(y,x,X,lam=0.1,niter=1,npm=1000)}

}
\references{
Chen, H. Y., Li, H., Argos, M., Persky, V. W., and Turyk, M. (2022). Statistical Methods
for Assessing Explained Variation of a Health Outcome by Mixture of Exposures. International Journal
of Environmental Research and Public Health.

Reference 2 to be added.
}
