% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eesdPMTca.R
\name{R2eesdPMTca}
\alias{R2eesdPMTca}
\title{Conditional permutation test for the estimating equation approach adjusting for covariates with supplementary data}
\usage{
R2eesdPMTca(y, x, X, pa, lam = 0.2, niter = 3, npm = 1000)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{X}{supplementary covariate data}

\item{pa}{dimension of covariates to be adjusted}

\item{lam}{parameter adjusting the format of the weighting matrix. Default is 0.2}

\item{niter}{number of iteration for updating lambda}

\item{npm}{Monte Carlo sample size for permutation}
}
\value{
Output of this function...
}
\description{
One paragraph describing this function
}
\details{
Details of this function...
}
\examples{
\dontrun{R2eesdPMTca(y,x,X,pa,lam=0.2,niter=3,npm=1000)}

}
\references{
Reference 1 here...

Reference 2 here...
}
