% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eesdPMTcaa.R
\name{R2eesdPMTcaa}
\alias{R2eesdPMTcaa}
\title{Conditional permutation test for no extra explained variation by estimating equation approach
with supplementary covariate data}
\usage{
R2eesdPMTcaa(y, x, X, pa, lam = 0.2, niter = 3, npm = 1000)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{X}{supplementary covariate data}

\item{pa}{dimension of covariates to be adjusted}

\item{lam}{parameter adjusting the format of the weighting matrix. Default is 0.2}

\item{niter}{number of iteration for updating lambda. Default is 3}

\item{npm}{Monte Carlo sample size for permutation. Default is 1000}
}
\value{
Output includes the p-values (estimate and bound) of the test,
estimate of proportion of the extra explained variation, and simulation results.
}
\description{
This function permutes the outcome and uses the estimator for the proportion of variation
explained for performing test of no explained variation.
}
\details{
The algorithm can be slow because the second part of the covariates need to be permuted.
The approach differs from the \code{R2eesdPMTca} in the way the matrix operations are handled.
}
\examples{
\dontrun{R2eesdPMTcaa(y, x, X, pa, lam = 0.1, niter = 3, npm = 1000)}

}
\references{
Chen, H.Y.; Li, H.; Argos, M.; Persky, V.; Turyk, M.
Statistical methods for assessing explained variations of a health outcome by mixtures of exposures.
Under review for Prep. Spec. Issue Int. J. Environ. Res. Public Health 2022.

An additional reference is to be added.
}
