% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2GCTA.R
\name{R2GCTA}
\alias{R2GCTA}
\title{GCTA approach}
\usage{
R2GCTA(y, x, lam, niter = 10, bt = "T", btn)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{lam}{parameter adjusting the format of the weighting matrix}

\item{niter}{number of iterations for updating lambda}

\item{bt}{Default is "T"}

\item{btn}{...}
}
\value{
Output of this function...
}
\description{
This function implements the GCTA approach assuming independent covariates and normally distributed errors, with bootstrap variance.
}
\details{
Details of this function...
}
\examples{
\dontrun{...}

}
\references{
Reference 1 here...

Reference 2 here...
}
