% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVee.R
\name{RVee}
\alias{RVee}
\title{Estimating equation approach to the proportion of the explained variation,
and the explained variation.}
\usage{
RVee(y, x, lam = 0.1, niter = 0, alpha = c(0.05))
}
\arguments{
\item{y}{outcome: a vector of length n.}

\item{x}{covariates: a matrix of nxp dimension.}

\item{lam}{parameter for altering the weighting matrix.}

\item{niter}{number of iterations for updating the lam parameter}

\item{alpha}{a vector of type I error for create the confidence intervals.}
}
\value{
Estimate of the proportion of explained variation, variance estimates, and confidence intervals,
under normality and non-normality assumptions.

Estimate of the explained variation, variance estimates, and confidence intervals
under normality and non-normality assumptions.
}
\description{
This function estimates:
(1). the proportion of the explained variation
(2). the explained variation
by covariates in a linear model assuming the covariates are independent.
}
\details{
Both point estimate and confidence intervals
are computed. Two set of confidence intervals under normal or non-normal error are computed.
}
\examples{
\dontrun{RVee(y,x)}

}
\references{
Chen, H.Y. (2022). Statistical inference on explained variation in high-dimensional
linear model with dense effects. arXiv:2201.08723

Chen, H. Y., Li, H., Argos, M., Persky, V. W., and Turyk, M. (2022). Statistical Methods
for Assessing Explained Variation of a Health Outcome by Mixture of Exposures. International Journal
of Environmental Research and Public Health.
}
