% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVeelsa.R
\name{RVeelsa}
\alias{RVeelsa}
\title{Estimating equation approach to the proportion of the explained variation using least square with
simulated variance estimate for inference (more accurate)}
\usage{
RVeelsa(
  y,
  x,
  lam = 0.2,
  niter = 1,
  alpha = 0.05,
  KV = rep(0, 3),
  know = "no",
  nrep = 1000
)
}
\arguments{
\item{y}{outcome: a vector of length n.}

\item{x}{covariates: a matrix of nxp dimension.}

\item{lam}{parameter for altering the weighting matrix.}

\item{niter}{number of iterations for updating lam.}

\item{alpha}{a vector of type I errors used to generate (1-alpha)confidence intervals.}

\item{KV, }{the first component of the vector KV=kappa_1, the second=kappa_2, the third=kappa_3}

\item{know}{if KV is known, options include "yes" and "no". Default is "no".}

\item{nrep}{Monte Carlo sample size for computinging KV.}
}
\value{
Estimate of the proportion of explained variation, variance estimates under normality
and non-normality assumptions, and confidence intervals under normality and non-normality assumptions.
}
\description{
This function computes:
(1). the proportion of the explained variation
(2). The explained variation
adjusting for the correlation in covariates for n>p
}
\details{
The estimation approach does not assume independent covariates and can deal
with the case \eqn{n\le p}. But require the sample sizes of x and X combined be greater than p.
}
\examples{
\dontrun{R2eelsa(y,x,lam=0.2,niter=3,alpha=c(0.1,0.05,0.01),KV=rep(0,3),know="no",nrep=1000)}


}
\references{
Chen, H. Y., Li, H., Argos, M., Persky, V. W., and Turyk, M. (2022). Statistical Methods
for Assessing Explained Variation of a Health Outcome by Mixture of Exposures. International Journal
of Environmental Research and Public Health.

Chen H. Y, Zhang, B. and Pan, G.(2023).Estimation and inference on explained variation
with possible supplementary data, Manuscript.
}
