% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVls.R
\name{RVls}
\alias{RVls}
\title{Least-square approach to proportion of the explained variation}
\usage{
RVls(y, x, lam = 1, alpha = c(0.05))
}
\arguments{
\item{y}{outcome: a vector of length n.}

\item{x}{covariates: a matrix of nxp dimension.}

\item{lam}{parameter for altering the weighting matrix.}

\item{alpha}{a vector of type I errors used to generate (1-alpha)confidence intervals.}
}
\value{
Estimate of proportion of the explained variation, variance estimates under normality
and non-normality error, and the corresponding confidence intervals.
}
\description{
This function uses least-square estimates in computing the proportion of the explained variation.
}
\details{
This method works only for the case n>p. It uses the least-square approach
for the estimation. Covariates are allowed to be correlated.
}
\examples{
\dontrun{RVls(y,x)}

}
\references{
Chen, H.Y. (2022). Statistical inference on explained variation in high-dimensional
linear model with dense effects. arXiv:2201.08723

Chen, H. Y., Li, H., Argos, M., Persky, V. W., and Turyk, M. (2022). Statistical Methods
for Assessing Explained Variation of a Health Outcome by Mixture of Exposures. International Journal
of Environmental Research and Public Health.
}
