% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVmldea.R
\name{RVmlea}
\alias{RVmlea}
\title{This function implement the maximum likelihood approach of Dicker and Erdogdu (2016).}
\usage{
RVmlea(y, x, alpha = c(0.05), lam = 1, niter = 100, eps = 1e-06)
}
\arguments{
\item{y}{outcome: a vector of length n.}

\item{x}{covariates: a matrix of nxp dimension.}

\item{alpha}{a vector of type I errors used to generate (1-alpha)confidence intervals.}

\item{lam}{initial value}

\item{niter}{the number of iterations for finding the signal noise ratio}

\item{eps}{the convergence criterion for the iteration}
}
\value{
Estimate of proportion of the explained variation, variance estimates,
and the corresponding confidence intervals.
}
\description{
This the mle for the high dimensional linear model
}
\details{
This method assume the independent covariates with fixed effects but can be
equivalently treated as random effects
}
\examples{
\dontrun{RVMLE(y,x)}

}
\references{
Dicker, L. H. and Erdogdu, M. A. (2016). Maximum likelihood for
variance estimation in high-dimensional linear models.
Proceedings of the 19th International Conference on Articial Intelligence and Statistics
}
