% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVsda.R
\name{RVsda}
\alias{RVsda}
\title{Estimating equation approach to the proportion of the explained variation with supplementary covariate data}
\usage{
RVsda(y, x, xsup, lam = 0.1)
}
\arguments{
\item{y}{outcome: a vector of length n.}

\item{x}{covariates: a matrix of nxp dimension.}

\item{xsup}{supplementary covariates: a matrix of Nxp dimension.
when xsup=NULL and n>p, it performs least-square with the simulation variance estimate}

\item{lam}{parameter for altering the weighting matrix.}
}
\value{
Estimate of the proportion of explained variation, variance estimates under normality
and non-normality assumptions, and confidence intervals under normality and non-normality assumptions.
}
\description{
This function computes:
(1). the proportion of the explained variation
(2). The explained variation
adjusting for the correlation in covariates with possible suplementary data
}
\details{
The estimation approach does not assume independent covariates and can deal
with the case \eqn{n\le p}. But require the sample sizes of x and X combined be greater than p.
}
\examples{
\dontrun{RVsda(y,x,xsup,lam=0.2)}


}
\references{
Chen, H. Y., Li, H., Argos, M., Persky, V. W., and Turyk, M. (2022). Statistical Methods
for Assessing Explained Variation of a Health Outcome by Mixture of Exposures. International Journal
of Environmental Research and Public Health.

Chen H. Y, Zhang, B. and Pan, G.(2023).Estimation and inference on explained variation
with possible supplementary data, Manuscript.
}
