% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVts.R
\name{RVts}
\alias{RVts}
\title{Estimating equation approach to the proportion of the explained variation,
and the explained variation with estimated structured covariance matrix}
\usage{
RVts(y, x, xsup = NULL, lam = 1, niter = 0, alpha = c(0.05), maxpc = 10)
}
\arguments{
\item{y}{outcome: a vector of length n.}

\item{x}{covariates: a matrix of nxp dimension.}

\item{xsup}{supplementary covariates (Nxp) if any.}

\item{lam}{parameter for altering the weighting matrix.}

\item{niter}{number of iterations for updating the lam parameter}

\item{alpha}{a vector of type I error for create the confidence intervals.}

\item{maxpc}{a number denotes the maximum number principal components allowed.}
}
\value{
Estimate of the proportion of explained variation, variance estimates, and confidence intervals,
under normality and non-normality assumptions.

Estimate of the explained variation, variance estimates, and confidence intervals
under normality and non-normality assumptions.
}
\description{
This function estimates:
(1). the proportion of the explained variation
(2). the explained variation
by covariates in a linear model assuming the covariates are correlated and the
covariance are estimated using the latent variable followed by
sparse precision matrix assumption.
}
\details{
Both point estimate and confidence intervals
are computed. Two set of confidence intervals under normal or non-normal error are computed.
}
\examples{
\dontrun{RVts(y,x)}

}
\references{
Chen, H.Y. (2026). Estimation of Proportion of Explained Variation by High-dimensional
covariates with Structured covariance Matrix.
}
