% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{sdgen}
\alias{sdgen}
\title{generate a square root of a correlation matrix}
\usage{
sdgen(amply, tilt, shrink, p)
}
\arguments{
\item{amply}{parameter controlling the correlation}

\item{tilt}{parameter controlling the correlation}

\item{shrink}{parameter(>=0) controlling the correlation}

\item{p}{dimensional of the matrix}
}
\value{
sqrtsig square root of a covariance matrix
}
\description{
This function generate data follows linear model based on parameter input.
}
