% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitDataAnalysis.R
\name{seldifdat}
\alias{seldifdat}
\title{Select data with index in dat1, but not in dat2. Assuming dat2 is part of dat1.}
\usage{
seldifdat(dat1, dat2, vind1, vind2)
}
\arguments{
\item{dat1}{The first data set to select data from}

\item{dat2}{The second data set to exclude data from data1 whose indices are in dat2.}

\item{vind1}{The index variable in dat1 used to determine the selection}

\item{vind2}{The index variable in dat2 used to exclude the data from dat1 whose indices match.
As such, vind1 in dat1 and vind2 in dat2 should be the same variable.}
}
\value{
Two data sets
Note that we can use merge function to select data with common index data.
}
\description{
Select data with index in dat1, but not in dat2. Assuming dat2 is part of dat1.
}
\details{
Separate dat1 into two data sets with one having indices of vind1 in dat1, but not in dat2,
and the other having indices match those of vind2 in dat2 ()
}
\examples{
\dontrun{seldifdat(dat1, dat2, vind1=1,vind2=1)}

}
