% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{simulation}
\alias{simulation}
\title{Perform simulation comparing methods including
EigenPrism
RVee
RVeels if (n>p)
RVeesd if supplementary data are available
projection+RVee}
\usage{
simulation(
  p = 200,
  n = 400,
  Nsd = 200,
  rept = 1000,
  nrep = 1000,
  cindep = "T",
  powx = 1,
  powy = 1,
  beta0 = 1.5,
  p1 = 100
)
}
\arguments{
\item{p}{exposure dimension}

\item{n}{sample size of original data}

\item{Nsd}{sample size of the supplementary covariates.}

\item{rept}{repetition of simulations}

\item{nrep}{Monte Carlo sample size for computinging KV.}

\item{cindep}{independet exposures or not}

\item{powx}{power of normal for x to simulation non-normal covariates}

\item{powy}{power of normal for y to simulation non-normal outcome}

\item{beta0}{non-zero regression coefficient}

\item{p1}{number of non-zero regression coefficients}
}
\value{
simulation results
}
\description{
Perform simulation comparing methods including
EigenPrism
RVee
RVeels if (n>p)
RVeesd if supplementary data are available
projection+RVee
}
\details{
perform simulation to compare different methods under varying situations
}
\examples{
\dontrun{simulation(p=200,n=400,Nsd=200, rept=1000,nrep=1000,
              cindep="T",powx=1,powy=1,beta0=1.5,p1=100)}

}
\references{
Chen, H. Y., Li, H., Argos, M., Persky, V. W., and Turyk, M. (2022). Statistical Methods
for Assessing Explained Variation of a Health Outcome by Mixture of Exposures. International Journal
of Environmental Research and Public Health.

Chen H. Y, Zhang, B. and Pan, G.(2023).Estimation and inference on explained variation
with possible supplementary data, Manuscript.
}
