% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Soutput}
\alias{Soutput}
\title{Output simulation results.}
\usage{
Soutput(result, pa, truth, rept)
}
\arguments{
\item{result}{a vector of simulation results}

\item{pa}{length of alpha (type I error vector)}

\item{truth}{true (proportion of) the explained variation}

\item{rept}{the number of simulations}
}
\value{
empirical confidence interval length and coverage.
}
\description{
This function summarize the simulation results with the input of simulation results.
}
