% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{sqrtpdm}
\alias{sqrtpdm}
\title{Find the square root of a positive definite matrix}
\usage{
sqrtpdm(corr)
}
\arguments{
\item{corr}{correlation or covariance matrix}
}
\value{
sqrtsig square root of a covariance matrix
}
\description{
This function find a square root of the positive definite matrix by single value decomposition.
}
