% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitDataAnalysis.R
\name{submitDataAnalysis}
\alias{submitDataAnalysis}
\title{Perform data analysis to estimate the proportion of explained variation from
outcome data: dataseta1; covariate data: dataseta2;
and the suplementary covariate data: dataseta3.}
\usage{
submitDataAnalysis(dataseta1, dataseta2, dataseta3, nimpute = 6, alpha = 0.05)
}
\arguments{
\item{dataseta1}{multiply imputed data set for outcomes and confounders.}

\item{dataseta2}{multiply imputed data set for covariates.}

\item{dataseta3}{multiply imputed data set for supplementary covariates.}

\item{nimpute}{copies of imputed data sets.}

\item{alpha}{to determine confidence as 1-alpha.}
}
\value{
estimators, variance estimates, and confidence intervals.
}
\description{
This function performance data analysis with multiply imputed datasets
}
\details{
perform both unadjusted and adjusted analyses and using Rubin's rule to combine results
}
\examples{
\dontrun{jasaDataAnalysis(outcome,covariates,supplecov)}

}
