% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transf}
\alias{transf}
\title{Decorrelation transformation}
\usage{
transf(x)
}
\arguments{
\item{x}{a matrix of nxp dimension.}
}
\value{
decorrelation data matrix.
}
\description{
This function performs a decorrelation transformation of a data set.
}
\details{
This function use the input data to estimate the covariance matrix,
and then use the estimated covariate matrix to decorrelate
the input data matrix. require n>p.
}
