% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{trueRV}
\alias{trueRV}
\title{Calculation of the true r2 and explained variance v2 by simulation}
\usage{
trueRV(nrep, p, beta, xsig, errsig, powx, powy, sd)
}
\arguments{
\item{nrep}{simulation sample size}

\item{p}{number of covariates}

\item{beta}{regression coefficients}

\item{xsig}{standard deviation of x variables}

\item{errsig}{residual standard deviation}

\item{powx}{power for x variable transformation}

\item{powy}{power of y variable transformation}

\item{sd}{square root decomposition of correlation matrix}
}
\value{
x covariates.
y outcome variable
}
\description{
Calculation of the true r2 and explained variance v2 by simulation
}
