% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zscale}
\alias{zscale}
\title{Transformation function}
\usage{
zscale(z)
}
\arguments{
\item{z}{input argument}
}
\value{
rescale data matrix.
}
\description{
This function rescales each variable in the data by subtracting mean and
divided by standard deviation estimates
}
