% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitBPs.R
\docType{data}
\name{BPs}
\alias{BPs}
\title{Bloop pressure data (outcomes)for the submitted paper}
\format{
a data frame with 22270 records and 3 variables.
}
\source{
NHANES plus imputation
}
\usage{
BPs
}
\description{
NHANES Data with 3 variables (ID, SBP, DBP), and 22270 observations
}
\keyword{datasets}
