% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenPrismFull.R
\name{EigenPrismFull}
\alias{EigenPrismFull}
\title{Extended EigenPrism approach}
\usage{
EigenPrismFull(y, x, alpha = c(0.01, 0.05, 0.1))
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{alpha}{significance level}
}
\value{
Output includes estimate of the proportion of the explained variation and the confidence intervals
}
\description{
This function combines the function from Jensen for n <= p and a chi-square inversion for confidence interval under
error variance for n > p.
}
\details{
For n > p, confidence intervals are obtained using least-square estimates.
}
\examples{
\dontrun{EigenPrismFull(y, x)}

}
\references{
Chen, H.Y. (2022). Statistical inference on explained variation in
high-dimensional linear model with dense effects. arXiv:2201.08723

Janson, L., Barber, R. F., Candes, E. (2017). EigenPrism: inference
for high-dimensional signal-to-noise ratios.
\emph{Journal of Royal Statistical Society, Ser. B.}, \strong{79}, 1037-1065.

Lucas Janson. \url{http://lucasjanson.fas.harvard.edu/code/EigenPrism.R}
}
