% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EigenPrismFull.R
\name{EigenPrismFull}
\alias{EigenPrismFull}
\title{Estimating proportion of explained variation using the least-square approach or the EigenPrism approach}
\usage{
EigenPrismFull(y, x, alpha = c(0.01, 0.05, 0.1))
}
\arguments{
\item{y}{outcome: a vector of length n.}

\item{x}{covariates: a matrix of nxp dimension.}

\item{alpha}{significance level: a vector.}
}
\value{
Estimate of the proportion of the explained variation and confidence intervals for the proportion.
}
\description{
EigenPrismFull procedure integrates the \eqn{n\le p} and \eqn{n>p} cases.
}
\details{
For \eqn{n\le p}, the estimate and confidence interval are obtained by EigenPrism approach.
For \eqn{n>p}, the estimate is obtained by least-square approach, and the confidence intervals
are obtained by inverting the chisquare test.
}
\examples{
\dontrun{EigenPrismFull(y, x)}

}
\references{
Chen, H.Y. (2022). Statistical inference on explained variation in
high-dimensional linear model with dense effects. arXiv:2201.08723.

Janson, L., Barber, R. F., Candes, E. (2017). EigenPrism: inference
for high-dimensional signal-to-noise ratios.
\emph{Journal of Royal Statistical Society, Ser. B.}, \strong{79}, 1037-1065.

Lucas Janson. \url{http://lucasjanson.fas.harvard.edu/code/EigenPrism.R}.
}
