% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{projed}
\alias{projed}
\title{Compute projection of x on the orthogonal complement of z}
\usage{
projed(x, z)
}
\arguments{
\item{x}{input matrix of nxp dimension}

\item{z}{input matrix of nxq dimension}
}
\value{
a nxp matrix, the projection of x on the orthogonal complement of z.
}
\description{
Compute projection of x on the orthogonal complement of z
}
