% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2ee.R
\name{R2ee}
\alias{R2ee}
\title{Computes the confidence interval with the estimating equation approach}
\usage{
R2ee(y, x, lam = 0.1, niter = 3)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{lam}{parameter adjusting the format of the weighting matrix. Default is 0.1}

\item{niter}{number of iterations for updating lambda}
}
\value{
The output is the confidence interval...
}
\description{
One paragraph describing this function...
}
\details{
details about this function...
}
\examples{
\dontrun{R2ee(y,x,lam=0.1,niter=3)}

}
\references{
reference 1 here...

reference 2 here...
}
