% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2ee.R
\name{R2ee}
\alias{R2ee}
\title{Proportion of the explained variation by the estimating equation approach}
\usage{
R2ee(y, x, lam = 0.1, niter = 3)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{lam}{initial parameter value (>= 0) for the weighting matrix. Default is 0.1}

\item{niter}{number of iterations for updating lam}
}
\value{
The output includes estimate of the proportion of explained variation, variance estimates
under normality and non-normality assumptions, and confidence intervals under normality and non-normality assumptions.
}
\description{
This function computes the estimator of the proportion of the explained variation assuming independent covariates.
}
\details{
The computation uses the singular value decomposition and asymptotic variance formula.
}
\examples{
\dontrun{R2ee(y, x, lam = 0.1,niter = 3)}

}
\references{
Chen, H.Y. (2022). Statistical inference on explained variation in high-dimensional
linear model with dense effects. arXiv:2201.08723
}
