% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eeadj.R
\name{R2eeadj}
\alias{R2eeadj}
\title{Estimating equation approach to the proportion of the explained variation with two subgroups.}
\usage{
R2eeadj(y, xa, xb, niter = 1)
}
\arguments{
\item{y}{outcome: a vector of length n.}

\item{xa}{covariates to be adjusted: a matrix of nxpa dimension.}

\item{xb}{covariate effects to be computed: a matrix of nxpb dimension.}

\item{niter}{number of iterations for updating lambda. Default is 1}
}
\value{
Estimates of the proportions of explained variations for xa, xb, and x=(xa,xb) respectively.
No variance estimate is available currently.
}
\description{
This approach estimates the proportion of the explained variation in a linear model
assuming the covariates are independent.
}
\details{
This method uses two estimating equations to estimate the variations of the outcome explained
by the two parts of covariates simultaneously.
}
\examples{
\dontrun{R2eeadj(y,xa,xb,niter=1)}

}
\references{
To be added.
}
