% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eeadj.R
\name{R2eeadj}
\alias{R2eeadj}
\title{Two-group estimating equation approach to proportion of the explained variation}
\usage{
R2eeadj(y, xa, xb, niter = 1)
}
\arguments{
\item{y}{outcome}

\item{xa}{covariates to be adjusted}

\item{xb}{covariate effects to be computed}

\item{niter}{number of iterations for updating lambda. Default is 1}
}
\value{
The output includes the estimates and confidence intervals.
}
\description{
This method uses two estimating equations to estimate the variations
of the outcome explained by the two parts of covariates simultaneously.
}
\details{
This method assumes covariates are independent.
}
\examples{
\dontrun{R2eeadj(y, xa, xb, niter = 1)}


}
\references{
To be added.
}
