% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eels.R
\name{R2eels}
\alias{R2eels}
\title{Confidence intervals for the least squares approach}
\usage{
R2eels(y, x)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}
}
\value{
Ooutput of this fucntion...
}
\description{
One paragraph describing this fucntion
}
\details{
Details of this function...
}
\examples{
\dontrun{R2eels(y,x)}

}
\references{
Reference 1 here...

Reference 2 here...
}
