% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eels.R
\name{R2eels}
\alias{R2eels}
\title{Least-square approach to proportion of the explained variation}
\usage{
R2eels(y, x)
}
\arguments{
\item{y}{outcome: a vector of length n.}

\item{x}{covariates: a matrix of nxp dimension.}
}
\value{
Estimate of proportion of the explained variation, variance estimates under normality
and non-normality error, and the corresponding confidence intervals.
}
\description{
This function uses least-square estimates in computing the proportion of the explained variation.
}
\details{
This method works only for the case n>p. It uses the least-square approach
for the estimation. Covariates are allowed to be correlated.
}
\examples{
\dontrun{R2eels(y,x)}

}
\references{
Chen, H.Y. (2022). Statistical inference on explained variation in high-dimensional
linear model with dense effects. arXiv:2201.08723.

Chen, H. Y., Li, H., Argos, M., Persky, V. W., and Turyk, M. (2022). Statistical Methods
for Assessing Explained Variation of a Health Outcome by Mixture of Exposures. International Journal
of Environmental Research and Public Health.
}
