% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eels.R
\name{R2eels}
\alias{R2eels}
\title{Estimate the proportion of explained variation by the least squares approach}
\usage{
R2eels(y, x)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}
}
\value{
Output includes estimates, variance, and confidence intervals.
}
\description{
This function works only for n > p. But covariates are not required to be independent.
}
\details{
This method uses the least square approach.
}
\examples{
\dontrun{R2eels(y,x)}

}
\references{
Chen, H.Y. (2022). Statistical inference on explained variation in high-dimensional
linear model with dense effects. arXiv:2201.08723
}
