% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eelsPMT.R
\name{R2eelsPMT}
\alias{R2eelsPMT}
\title{Permutation test for no explained variation using least-square estimate}
\usage{
R2eelsPMT(y, x, npm = 1000)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{npm}{permutation sample size. Default is 1000}
}
\value{
Output includes the p-values (estimate and bound) of the test,
estimate of proportion of the explained variation, and simulation results.
}
\description{
This function performs test of no explained variation by
permuting the outcome using least square approach.
}
\details{
The computation permutes outcome and is computationally fast.
}
\examples{
\dontrun{R2eelsPMT(y, x, npm=1000)}

}
\references{
Chen, H.Y.; Li, H.; Argos, M.; Persky, V.; Turyk, M.
Statistical methods for assessing explained variations of a health outcome by mixtures of exposures.
Under review for Prep. Spec. Issue Int. J. Environ. Res. Public Health 2022.
}
