% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eelsPMT.R
\name{R2eelsPMT}
\alias{R2eelsPMT}
\title{Permutation test for the least squares approach}
\usage{
R2eelsPMT(y, x, npm = 1000)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{npm}{permutation sample size. Default is 1000}
}
\value{
Ouput of this function...
}
\description{
One paragraph describing this function
}
\details{
Details of this function...
}
\examples{
\dontrun{R2eelsPMT(y,x,npm=1000)}

}
\references{
Reference 1 here...

Reference 2 here...
}
