% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eelsPMTcaa.R
\name{R2eelsPMTcaa}
\alias{R2eelsPMTcaa}
\title{Conditional permutation test for no explained variation using the least-square approach}
\usage{
R2eelsPMTcaa(y, x, pa, npm = 1000)
}
\arguments{
\item{y}{outcome}

\item{x}{all the covariates}

\item{pa}{dimension of the first part of covariates to be adjusted}

\item{npm}{permutation sample size. Default is 1000}
}
\value{
Output includes the p-values (estimate and bound) of the test,
estimate of proportion of the extra explained variation, and simulation results.
}
\description{
This function performs conditional tests of no variation explained by the additional
covariates when adjusted for the one set of covariates.
}
\details{
The conditional permutation test needs to perform the second part of the covariates
and can be computationally slow. This function differs from \code{R2eelsPMTca} in the way the matrix operation is used.
}
\examples{
\dontrun{R2eelsPMTcaa(y, x, pa, npm = 1000)}

}
\references{
Chen, H.Y.; Li, H.; Argos, M.; Persky, V.; Turyk, M.
Statistical methods for assessing explained variations of a health outcome by mixtures of exposures.
Under review for Prep. Spec. Issue Int. J. Environ. Res. Public Health 2022.
}
