% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eePMT.R
\name{R2eePMT}
\alias{R2eePMT}
\title{Permutation test for no explained variation}
\usage{
R2eePMT(y, x, lam = 0.12, niter = 3, npm = 1000)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{lam}{parameter adjusting the formation of the weighting matrix. Default is 0.12}

\item{niter}{number of iterations for updating lambda. Default is 3}

\item{npm}{permutation sample size. Default is 1000}
}
\value{
Output includes the p-values (estimate and bound) of the test,
estimate of proportion of the explained variation, and simulation results.
}
\description{
This function performs test of no explained
variation by permuting the outcome using the estimating equation approach.
}
\details{
The computation permutes outcome and can be computationally fast.
}
\examples{
\dontrun{R2eePMT(y, x, lam = 0.12, niter = 3, npm = 1000)}

}
\references{
Chen, H.Y.; Li, H.; Argos, M.; Persky, V.; Turyk, M.
Statistical methods for assessing explained variations of a health outcome by mixtures of exposures.
Under review for Prep. Spec. Issue Int. J. Environ. Res. Public Health 2022.
}
