% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eePMTca.R
\name{R2eePMTca}
\alias{R2eePMTca}
\title{Conditional permutation test adjusting for covariates for estimating equation approach}
\usage{
R2eePMTca(y, x, pa, lam = 0.2, niter = 3, npm = 1000)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{pa}{dimension of covariates to be adjusted}

\item{lam}{parameter adjusting the format of the weighting matrix. Default is 0.2}

\item{niter}{number of iteration for updating lambda. Default is 3}

\item{npm}{Monte Carlo sample size for permutation. Default is 1000}
}
\value{
Output includes the p-values (estimate and bound) of the test,
estimate of proportion of the extra explained variation, and simulation results.
}
\description{
This function performs test of no extra variation explained by the additional covariates
by permuting the additional covariates using the estimating equation approach.
}
\details{
The computation permutes the second part of the covariates and is computationally slow.
}
\examples{
\dontrun{R2eePMTca(y, x, pa, lam = 0.2, niter = 3, npm = 1000)}

}
\references{
Chen, H.Y.; Li, H.; Argos, M.; Persky, V.; Turyk, M.
Statistical methods for assessing explained variations of a health outcome by mixtures of exposures.
Under review for Prep. Spec. Issue Int. J. Environ. Res. Public Health 2022.
}
