% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eePMTcaa.R
\name{R2eePMTcaa}
\alias{R2eePMTcaa}
\title{Conditional permutation test adjusting for covariates for estimating equation approach}
\usage{
R2eePMTcaa(y, x, pa, lam = 0.2, niter = 3, npm = 1000)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{pa}{dimension of covariates to be adjusted}

\item{lam}{parameter adjusting the format of the weighting matrix. Default is 0.2}

\item{niter}{number of iterations for updating lambda. Default is 3}

\item{npm}{Monte Carlo sample size for permutation. Default is 1000}
}
\value{
Output is the estimate of the proportion of the variation explained by
the additional covariates and the p-value of the test.
}
\description{
This function performs test of no extra variation explained by the additional covariates
by permuting the additional covariates using the estimating equation approach.
}
\details{
The computation permutes the second part of the covariates and is computationally slow.
It differs from \code{R2eePMTca} in the way of matrix computation.
}
\examples{
\dontrun{R2eePMTcaa(y,x,pa,lam=0.2,niter=3,npm=1000)}

}
\references{
Chen, H.Y.; Li, H.; Argos, M.; Persky, V.; Turyk, M.
Statistical methods for assessing explained variations of a health outcome by mixtures of exposures.
Prep. Spec. Issue Int. J. Environ. Res. Public Health 2022.
}
