% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eesd.R
\name{R2eesd}
\alias{R2eesd}
\title{Estimation and inference on the explained variation by the estimating equation approach
with supplementary covariate data}
\usage{
R2eesd(
  y,
  x,
  X,
  lam = 0.2,
  niter = 3,
  V = rep(0, 3),
  E = rep(0, 3),
  know = "yes",
  nrep = 1000
)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{X}{supplementary covariate data}

\item{lam}{parameter adjusting the format of the weighting matrix. Default is 0.2}

\item{niter}{number of iterations for updating lambda. Default is 3}

\item{V}{\code{V[1:3] == V[1]=VA}, \code{V[2]=VAB}, \code{V[3]=VB}}

\item{E}{\code{E[1:3] == E[1]=EC}, \code{E[2]=ED}, \code{E[3]=EF}}

\item{know}{if V and E are known, options include "yes" and "no". Default is "yes"}

\item{nrep}{Monte Carlo sample size for updating V and E}
}
\value{
Output includes the estimator of the proportion of variation explained and
its variance, and the confidence intervals.
}
\description{
This function estimates the proportion of the explained variation using the estimating
equation approach incorporating supplementary covariate data.
}
\details{
The estimation approach does not assume independent covariates and can
deal with the case n <= p.
}
\examples{
\dontrun{R2eesd(y,x,X,lam=0.2,niter=3,V=rep(0,3),E=rep(0,3),know="yes",nrep=1000)}


}
\references{
Chen, H.Y.; Li, H.; Argos, M.; Persky, V.; Turyk, M.
Statistical methods for assessing explained variations of a health outcome by mixtures of exposures.
Prep. Spec. Issue Int. J. Environ. Res. Public Health 2022.

An additional reference is to be added.
}
