% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eesd.R
\name{R2eesd}
\alias{R2eesd}
\title{Estimating equation approach to the proportion of the explained variation with supplementary covariate data}
\usage{
R2eesd(
  y,
  x,
  X,
  lam = 0.2,
  niter = 3,
  V = rep(0, 3),
  E = rep(0, 3),
  know = "yes",
  nrep = 1000
)
}
\arguments{
\item{y}{outcome: a vector of length n.}

\item{x}{covariates: a matrix of nxp dimension.}

\item{X}{supplementary covariates: a matrix of Nxp dimension.}

\item{lam}{parameter for altering the weighting matrix.}

\item{niter}{number of iterations for updating lam.}

\item{V}{\code{V[1:3] == V[1]=VA, V[2]=VAB, V[3]=VB}}

\item{E}{\code{E[1:3] == E[1]=EC, E[2]=ED, E[3]=EF}}

\item{know}{if VA and EB are known, options include "yes" and "no". Default is "yes" Default is "yes" (usually for simulation only).}

\item{nrep}{Monte Carlo sample size for computing V and E.}
}
\value{
Estimate of the proportion of explained variation, variance estimates under normality
and non-normality assumptions, and confidence intervals under normality and non-normality assumptions.
}
\description{
This function computes the proportion of the explained variation adjusting for the correlation in covariates.
}
\details{
The estimation approach does not assume independent covariates and can deal
with the case \eqn{n\le p}. But require the sample sizes of x and X combined be greater than p.
}
\examples{
\dontrun{R2eesd(y,x,X,lam=0.2,niter=3,V=rep(0,3),E=rep(0,3),know="no",nrep=1000)}

}
\references{
Chen, H. Y., Li, H., Argos, M., Persky, V. W., and Turyk, M. (2022). Statistical Methods
for Assessing Explained Variation of a Health Outcome by Mixture of Exposures. International Journal
of Environmental Research and Public Health.

reference 2 to be added.
}
