% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eesd0.R
\name{R2eesd0}
\alias{R2eesd0}
\title{Funtion for temporary comparison}
\usage{
R2eesd0(y, x, X, lam, niter = 1, VA = 0, EB = 0, know = "yes", nrep = 1000)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{X}{supplementary covariate data}

\item{lam}{parameter adjusting the format of the weighting matrix}

\item{niter}{number of iterations for updating lambda}

\item{VA}{the first item of the variance value if known}

\item{EB}{the second term of the variance expression if known}

\item{know}{"yes" if VA and EB are known, "no" otherwise}

\item{nrep}{Monte Carlp sample size for computing VA and EB}
}
\description{
Funtion for temporary comparison
}
