% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eesdPMT.R
\name{R2eesdPMT}
\alias{R2eesdPMT}
\title{Permutation test for estimating equation approach with supplementary covariate data}
\usage{
R2eesdPMT(y, x, X, lam, niter = 3, npm = 1000)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{X}{supplementary covariate data}

\item{lam}{parameter adjusting the format of the weighting matrix}

\item{niter}{number of iterations for updating lambda. Default is 3}

\item{npm}{Monte Carlo sample size for permutation. Default is 1000}
}
\value{
Output of this function...
}
\description{
One paragraph describing this function...
}
\details{
Details of this function...
}
\examples{
\dontrun{R2eesdPMT(y,x,X,lam,niter=3,npm=1000)}

}
\references{
Reference 1 here...

Reference 2 here...
}
