% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eesdPMT.R
\name{R2eesdPMT}
\alias{R2eesdPMT}
\title{Permutation test for no explained variation by
estimating equation approach with supplementary covariate data}
\usage{
R2eesdPMT(y, x, X, lam, niter = 3, npm = 1000)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{X}{supplementary covariate data}

\item{lam}{parameter adjusting the format of the weighting matrix}

\item{niter}{number of iterations for updating lambda. Default is 3}

\item{npm}{Monte Carlo sample size for permutation. Default is 1000}
}
\value{
Output is the estimator of the proportion of explained variation
and the p-value for the test.
}
\description{
This function permutes the outcome and use the estimator for the
proportion of variation explained for performing test of no explained variation.
}
\details{
The algorithm is fast because only the outcome is permuted.
}
\examples{
\dontrun{R2eesdPMT(y,x,X,lam,niter=3,npm=1000)}

}
\references{
Chen, H.Y.; Li, H.; Argos, M.; Persky, V.; Turyk, M.
Statistical methods for assessing explained variations of a health outcome by mixtures of exposures.
Prep. Spec. Issue Int. J. Environ. Res. Public Health 2022.

An additional reference is to be added..
}
