% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eesdPMTca.R
\name{R2eesdPMTca}
\alias{R2eesdPMTca}
\title{Permutation test for no additional explained variation by adding extra
covariates using the estimating equation approach incoporating supplementary covariate data}
\usage{
R2eesdPMTca(y, x, X, pa, lam = 0.2, niter = 3, npm = 1000)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{X}{supplementary covariate data}

\item{pa}{dimension of covariates to be adjusted}

\item{lam}{parameter adjusting the format of the weighting matrix. Default is 0.2}

\item{niter}{number of iteration for updating lambda}

\item{npm}{Monte Carlo sample size for permutation}
}
\value{
Output is the estimator of the proportion of explained variation and the p-value for the test.
}
\description{
This function permutes the outcome and use the estimator for the proportion of
variation explained for performing test of no explained variation.
}
\details{
The algorithm can be slow because the second part of the covariates need to be permuted.
}
\examples{
\dontrun{R2eesdPMTca(y,x,X,pa,lam=0.2,niter=3,npm=1000)}

}
\references{
Chen, H.Y.; Li, H.; Argos, M.; Persky, V.; Turyk, M.
Statistical methods for assessing explained variations of a health outcome by mixtures of exposures.
Prep. Spec. Issue Int. J. Environ. Res. Public Health 2022.

An additional reference is to be added.
}
