% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2eesdPMTcaa.R
\name{R2eesdPMTcaa}
\alias{R2eesdPMTcaa}
\title{Another conditional permutation test...}
\usage{
R2eesdPMTcaa(y, x, X, pa, lam = 0.1, niter = 3, npm = 1000)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{X}{supplementary covariate data}

\item{pa}{dimension of covariates to be adjusted}

\item{lam}{parameter adjusting the format of the weighting matrix. Default is 0.1}

\item{niter}{number of iteration for updating lambda. Default is 3}

\item{npm}{Monte Carlo sample size for permutation. Default is 1000}
}
\value{
Ouput of this function...
}
\description{
One paragraph describing this function...
}
\details{
Details of this function...
}
\examples{
\dontrun{R2eesdPMTcaa(y,x,X,pa,lam=0.1,niter=3,npm=1000)}

}
\references{
Reference 1 here...

Reference 2 here...
}
