% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2GCTA.R
\name{R2GCTA}
\alias{R2GCTA}
\title{The GCTA approach with bootstrap variance}
\usage{
R2GCTA(y, x, lam, niter = 10, bt = "T", btn)
}
\arguments{
\item{y}{outcome}

\item{x}{covariates}

\item{lam}{parameter adjusting the format of the weighting matrix}

\item{niter}{number of iterations for updating lambda}

\item{bt}{default is "T"}

\item{btn}{bootstrap sample size}
}
\value{
Output includes estimate of proportion of the explained variation,
and bootstrap estimate, variance, and confidence intervals if bt = T.
}
\description{
This function implements the GCTA approach assuming independent covariates
and normally distributed errors, with bootstrap variance.
}
\details{
The function uses the singular value decomposition for estimation and bootstrap
sampling approach for variance estimation under normal random errors.
}
\examples{
\dontrun{R2GCTA(y, x)}

}
\references{
Schweiger, R., Kaufman, S., Laaksonen, R., Kleber, M. E., Marz, W., Eskin, E., Rosset, S., Halperin, E. (2016).
Fats and ac-curate construction of confidence intervals for heritability.
\emph{The American Journal of Human Genetics}, \strong{98}, 1181-1192.

Yang, J., Lee, S. H., Wray, N. R., Goddard, M. E., Visscher, P. (2016).
GCTA-GREML accounts for linkage disequilibrium when estimating genetic variance from genome-wide SNPs.
\emph{Proceedings of the National Academy of Sciences}, \strong{113}, E4579-E4580.
}
