% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soutput.R
\name{Soutput}
\alias{Soutput}
\title{Output simulation results.}
\usage{
Soutput(result, r20, rept)
}
\arguments{
\item{result}{a vector of simulation results}

\item{r20}{true proportion of the explained variation}

\item{rept}{the number of simulations}
}
\value{
empirical confidence interval length and coverage.
}
\description{
This function summarize the simulation results with the input of simulation results.
}
