% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transfsd}
\alias{transfsd}
\title{Decorrelation transformation with supplementary data}
\usage{
transfsd(x, X)
}
\arguments{
\item{x}{input 1}

\item{X}{input 2}
}
\value{
Output is the decorrelation matrix for x.
}
\description{
Decorrelation with the aid of supplementary data
}
