% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zscale}
\alias{zscale}
\title{Rescale data matrix}
\usage{
zscale(z)
}
\arguments{
\item{z}{input matrix of nxp dimension.}
}
\value{
rescaled data matrix of nxp dimension.
}
\description{
Rescale data matrix by subtracting mean and divide the standard deviation for each variable.
}
